file manager now reads input from the textfile
