+++ 5001Project/Assets/Scripts/FileManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3e5e2b9b7cffd9b4eb3f8bdbbf741a11
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
