+++ 5001Project/Assets/Sprites/icon.mdp.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b1b64d3a7aedd4843b199fde6c51ea04
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
